+++ Kits/rokt-sdk-plus/rokt-sdk-plus-ios/Package.swift
@@ -3,7 +3,7 @@
 import Foundation
 import PackageDescription
 
-let version = "9.3.2"
+let version = "9.3.3"
 
 let useLocalVersion = ProcessInfo.processInfo.environment["USE_LOCAL_VERSION"] != nil